requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
Markdown==3.4.3
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
